backend/requirements.txt
google-cloud-bigquery
pyarrow
ipykernel
pandas
numpy
fastapi==0.108.0
pytz
uvicorn
colorama
db-dtypes
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
scipy==1.10.0
scikit-learn==1.3.1
mlflow==2.1.1
prefect==2.19.2
python-dotenv
psycopg2-binary
yfinance
pushover
pytest
requests-mock
datetime
mlflow
seaborn
statsmodels
click
